+++ pykechain/models/value_filter.py
@@ -3,8 +3,15 @@ import urllib
 import warnings
 from abc import abstractmethod
 from typing import Any, Dict, List, Optional, Union
+from urllib.parse import unquote
 
-from pykechain.enums import Category, FilterType, PropertyType, ScopeStatus
+from pykechain.enums import (
+    Category,
+    ContextGroup,
+    FilterType,
+    PropertyType,
+    ScopeStatus,
+)
 from pykechain.exceptions import IllegalArgumentError, NotFoundError
 from pykechain.models.input_checks import (
     check_base,
@@ -75,7 +82,7 @@ class PropertyValueFilter(BaseFilter):
 
         self.id = property_model_id
         if isinstance(value, str):
-            self.value = urllib.parse.unquote(value)
+            self.value = unquote(value)
         else:
             self.value = value
         self.type = filter_type
@@ -113,19 +120,17 @@ class PropertyValueFilter(BaseFilter):
 
         if prop.category != Category.MODEL:
             raise IllegalArgumentError(
-                'Property value filters can only be set on Property models, received "{}".'.format(
-                    prop
-                )
+                f'Property value filters can only be set on Property models, received "{prop}".'
             )
         else:
             property_type = prop.type
             if (
                 property_type
                 in (
-                    PropertyType.BOOLEAN_VALUE,
-                    PropertyType.REFERENCES_VALUE,
-                    PropertyType.ACTIVITY_REFERENCES_VALUE,
-                )
+                PropertyType.BOOLEAN_VALUE,
+                PropertyType.REFERENCES_VALUE,
+                PropertyType.ACTIVITY_REFERENCES_VALUE,
+            )
                 and self.type != FilterType.EXACT
             ):
                 warnings.warn(
@@ -190,9 +195,11 @@ class PropertyValueFilter(BaseFilter):
     @classmethod
     def parse_options(cls, options: Dict) -> List["PropertyValueFilter"]:
         """
-        Convert the dict & string-based definition of a property value filter to a list of PropertyValueFilter objects.
+        Convert the dict & string-based definition of a property value filter to a list of
+        PropertyValueFilter objects.
 
-        :param options: options dict from a multi-reference property or meta dict from a filtered grid widget.
+        :param options: options dict from a multi-reference property or meta dict from a filtered
+            grid widget.
         :return: list of PropertyValueFilter objects
         :rtype list
         """
@@ -235,7 +242,8 @@ class ScopeFilter(BaseFilter):
     :ivar tag: string
     """
 
-    # map between KE-chain field and Pykechain attribute, and whether the filter is stored as a list (cs-string)
+    # map between KE-chain field and Pykechain attribute, and whether the filter is stored as a
+    # list (cs-string)
     MAP = [
         ("name__icontains", "name", False),
         ("status__in", "status", False),
@@ -325,9 +333,10 @@ class ScopeFilter(BaseFilter):
     @classmethod
     def parse_options(cls, options: Dict) -> List["ScopeFilter"]:
         """
-        Convert the dict & string-based definition of a scope filter to a list of ScopeFilter objects.
+        Convert the dict & string-based definition of a scope filter to a list of ScopeFilter obj.
 
-        :param options: options dict from a scope reference property or meta dict from a scopes widget.
+        :param options: options dict from a scope reference property or meta dict from a scopes
+            widget.
         :return: list of ScopeFilter objects
         :rtype list
         """
@@ -377,8 +386,8 @@ class ScopeFilter(BaseFilter):
 
                 if filter_value is not None:
                     if is_list:
-                        # creata a string with commaseparted prefilters, the first item directly and
-                        # consequent items with a ,
+                        # creata a string with commaseparted prefilters, the first item directly
+                        # and consequent items with a ,
                         # TODO: refactor to create a list and then join them with a ','
                         if field not in prefilters:
                             prefilters[field] = filter_value
@@ -392,5 +401,68 @@ class ScopeFilter(BaseFilter):
 
             if not found:
                 prefilters.update(f.extra_filter)
+
+        return options
+
+
+class ContextgroupFilter(BaseFilter):
+    """A representation object for a ContextGroup filter.
+
+    There can only be a single context_group prefilter. So no  lists involved.
+
+    :ivar value: the value of the context_group filter, containing a
+        ContextGroup enumeration
+    """
+
+    def __init__(
+        self,
+        context_group: Union[str, ContextGroup],
+    ):
+        """Create ContextgroupFilter instance.
+
+        :var context_group: the value of the context_group filter, containing a
+            ContextGroup enumeration
+        """
+        check_enum(context_group, ContextGroup, "context_group")
+        self.value: ContextGroup = context_group
+
+    @classmethod
+    def parse_options(cls, options: Dict) -> "ContextgroupFilter":
+        """
+        Convert the dict definition of a context_group filter to a list of ContextgroupFilter obj.
+
+        The value_options of the context reference properties looks like:
+
+        ```
+        {
+            "prefilters": {
+                "context_group": "DISCIPLINE"
+            },
+        }
+        ```
+
+        :param options: options dict from a scope reference property or meta dict from a scopes
+            widget.
+        :return: list of ScopeFilter objects
+        :rtype list
+        """
+        filters_dict = options.get(MetaWidget.PREFILTERS, {})
+        for field, value in filters_dict.items():
+            if field == "context_group":
+                return cls(context_group=value)
+
+    @classmethod
+    def write_options(cls, filter: "ContextgroupFilter") -> Dict:
+        """
+        Convert the list of Filter objects to a dict.
+
+        :param filters: List of BaseFilter objects
+        :returns options dict to be used to update the options dict of a property
+        """
+        prefilters = dict()
+        options = {MetaWidget.PREFILTERS: prefilters}
+
+        if filter.value in ContextGroup:
+            prefilters.update({"context_group": str(filter.value)})
 
         return options